+++ creme/ensemble/__init__.py
@@ -4,8 +4,8 @@ A module for ensemble learning.
 from .bagging import BaggingClassifier
 from .bagging import BaggingRegressor
 from .grouping import GroupRegressor
-from .hedging import HedgeBinaryClassifier
 from .hedging import HedgeRegressor
+from .majority import WeightedMajorityClassifier
 from .stacking import StackingBinaryClassifier
 
 
@@ -15,5 +15,6 @@ __all__ = [
     'GroupRegressor',
     'HedgeBinaryClassifier',
     'HedgeRegressor',
-    'StackingBinaryClassifier'
+    'StackingBinaryClassifier',
+    'WeightedMajorityClassifier'
 ]