+++ CHANGELOG.md
@@ -17,6 +17,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - `IQR` and `RollingIQR` to `stats`
 - The `initializers` submodule has been added to the `optim` module, it can be used for initializing weights in weight-based models such as linear regression
 - `shuffle` to `stream`
+- `WeightedMajorityClassifier` to `ensemble`
 
 ### Modified
 
@@ -30,6 +31,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Renamed `stream.iter_numpy` to `stream.iter_array`
 - The module `optim` has been reorganized into submodules; namely `schedulers`, `initializers`, and `losses`. The top-level now only contains optimizers. Some classes have renamed accordingly. See the documentation for details.
 - `stream.iter_csv` can now read from `gzip` files
+
+### Removed
+
+- `HedgeBinaryClassifier` from `ensemble`, as it's performance was subpar
 
 ## [0.3.0](https://pypi.org/project/creme/0.3.0/) - 2019-06-23